requirements.txt
Flask==0.10.1
Flask-SocketIO==0.5.0
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.5
itsdangerous==0.24
pyserial==2.7
wsgiref==0.1.2
